requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Mail==0.10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
oauthlib==3.2.2
psycopg2-binary==2.9.9
requests==2.32.3
requests-oauthlib==2.0.0
urllib3==2.2.3
Werkzeug==3.0.4
